requirements.txt
Flask==2.3.3
tensorflow==2.17.0
numpy==1.24.3
pillow==10.0.0 
distutils
